haproxy-config.template: Use power-of-two balancing
Change the default balancing algorithm for insecure, edge-terminated, and
reencrypt routes to "random".

HAProxy 2.0 changes the "random" balancing algorithm to use 2 draws by
default, which means the algorithm is effectively the "Power of Two Random
Choices" algorithm.  If a route has only 1 or 2 endpoints, power-of-two
degenerates to leastconn (the default balancing algorithm before this
commit).  If a route has more than 2 endpoints, power-of-two tends to
approximate leastconn, and can perform better in scenarios where balancing
decision are decentralized (as is the case when multiple routers are
balancing to the same set of endpoints) by reducing the likelihood that all
HAProxy instances choose the same endpoint.

See <https://www.haproxy.com/blog/power-of-two-load-balancing/> for
performance comparisons of the random-with-1-draw, roundrobin,
power-of-two, and leastconn algorithms.

* images/router/haproxy/conf/haproxy-config.template: Modify the
$balanceAlgoPattern variable to match "random".  Change the default
balancing algorithm for insecure, edge-terminated, and reencrypt routes to
"random".