Read configuration from filesystem
If the conf/environ/ directory exists, watch it for updates, and update the
openshift-router process's environment when updates within conf/environ/
are observed.

For environment variables that are used by the controller, re-exec the
process so as to re-initialize the controller with the updated values if
any of those variables' values change.  Otherwise, for environment
variables that are only used in the template itself, force an HAProxy
reload if any of those variables' values change.

This commit is related to NE-1065.

https://issues.redhat.com/browse/NE-1065

* pkg/router/template/router.go (environDir): New const for the directory
that holds the environment variables.
(newTemplateRouter): Call the new watchEnvironDir function.
(templateOnlyEnvVar): New map of environment variables that do not require
a re-exec when changed.
(watchEnvironDir): New function to watch the environ directory, update the
openshift-router process's environment, and reload or re-exec as needed.